Refactor code and add docs
