+++ Sources/Uno/Extensions/Data+Extensions.swift
@@ -1,9 +1,11 @@
 import Foundation
 
 public extension Data {
-  /// An hexadecimal representation of the packet.
-  func hexString(enablePadding: Bool = true) -> String {
-    let hexDigitFormat = enablePadding ? "02" : ""
+  /// An hexadecimal representation of the bytes.
+  /// - Parameter forcingZeroPadding: Whether or not leading zeroes should be added to the string.
+  /// - Returns: A `String` representing the bytes sequence in hexadecimal format.
+  func hexString(forcingZeroPadding shouldForcePadding: Bool = true) -> String {
+    let hexDigitFormat = shouldForcePadding ? "02" : ""
     return map { String(format: "%\(hexDigitFormat)hhx", $0) }.joined()
   }
 }
